chore: move falling logic to its own namespace
