+++ src/MakaMek.Core/Models/Game/Mechanics/Mechs/Falling/IPilotingSkillCalculator.cs
@@ -2,7 +2,7 @@ using Sanet.MakaMek.Core.Data.Game.Mechanics;
 using Sanet.MakaMek.Core.Models.Map;
 using Sanet.MakaMek.Core.Models.Units;
 
-namespace Sanet.MakaMek.Core.Models.Game.Mechanics;
+namespace Sanet.MakaMek.Core.Models.Game.Mechanics.Mechs.Falling;
 
 /// <summary>
 /// Interface for calculating piloting skill roll target numbers
