+++ src/MakaMek.Core/Models/Game/Mechanics/Mechs/Falling/PilotingSkillCalculator.cs
@@ -7,7 +7,7 @@ using Sanet.MakaMek.Core.Models.Units.Components.Internal;
 using Sanet.MakaMek.Core.Models.Units.Mechs;
 using Sanet.MakaMek.Core.Utils.TechRules;
 
-namespace Sanet.MakaMek.Core.Models.Game.Mechanics;
+namespace Sanet.MakaMek.Core.Models.Game.Mechanics.Mechs.Falling;
 
 /// <summary>
 /// Classic BattleTech implementation of piloting skill roll calculator
