+++ tests/MakaMek.Core.Tests/Models/Game/Mechanics/Mechs/Falling/FallingDamageCalculatorTests.cs
@@ -1,6 +1,6 @@
 using NSubstitute;
 using Sanet.MakaMek.Core.Models.Game.Dice;
-using Sanet.MakaMek.Core.Models.Game.Mechanics;
+using Sanet.MakaMek.Core.Models.Game.Mechanics.Mechs.Falling;
 using Sanet.MakaMek.Core.Models.Map;
 using Sanet.MakaMek.Core.Models.Units;
 using Sanet.MakaMek.Core.Models.Units.Components.Engines;
@@ -9,7 +9,7 @@ using Sanet.MakaMek.Core.Tests.Models.Units;
 using Sanet.MakaMek.Core.Utils.TechRules;
 using Shouldly;
 
-namespace Sanet.MakaMek.Core.Tests.Models.Game.Mechanics;
+namespace Sanet.MakaMek.Core.Tests.Models.Game.Mechanics.Mechs.Falling;
 
 public class FallingDamageCalculatorTests
 {
